+++ src/index.ts
@@ -2,6 +2,6 @@ export { defineBddConfig, defineBddProject } from './config';
 export { BDDInputConfig } from './config/types';
 export { createBdd } from './steps/createBdd';
 export { test } from './run/bddFixtures/test';
-export { cucumberReporter } from './reporter/cucumber/helper';
+export { cucumberReporter } from './reporter/cucumber/wrapper';
 export { defineParameterType, IParameterTypeDefinition } from './steps/parameterTypes';
 export { DataTable } from './cucumber/DataTable';