+++ src/reporter/cucumber/base.ts
@@ -8,6 +8,7 @@ import fs from 'node:fs';
 import { Writable } from 'node:stream';
 import { finished } from 'node:stream/promises';
 import { EventEmitter } from 'node:events';
+import * as messages from '@cucumber/messages';
 import EventDataCollector from '../../cucumber/formatter/EventDataCollector.js';
 
 export type InternalOptions = {
@@ -15,6 +16,8 @@ export type InternalOptions = {
   eventBroadcaster: EventEmitter;
   eventDataCollector: EventDataCollector;
 };
+
+export type SkipAttachments = boolean | ('image/png' | 'video/webm' | 'application/zip' | string)[];
 
 export default class BaseReporter {
   protected outputStream: Writable = process.stdout;
@@ -52,4 +55,14 @@ export default class BaseReporter {
 
 function isStdout(stream: Writable) {
   return stream === process.stdout;
+}
+
+export function isAttachmentAllowed(
+  envelope: messages.Envelope,
+  skipAttachments?: SkipAttachments,
+) {
+  if (!('attachment' in envelope)) return true;
+  return Array.isArray(skipAttachments)
+    ? !skipAttachments.includes(envelope.attachment?.mediaType || '')
+    : skipAttachments !== true;
 }