+++ src/reporter/cucumber/message.ts
@@ -3,10 +3,11 @@
  * Based on: https://github.com/cucumber/cucumber-js/blob/main/src/formatter/message_formatter.ts
  */
 import * as messages from '@cucumber/messages';
-import BaseReporter, { InternalOptions } from './base';
+import BaseReporter, { InternalOptions, isAttachmentAllowed, SkipAttachments } from './base';
 
 type MessageReporterOptions = {
   outputFile?: string;
+  skipAttachments?: SkipAttachments;
 };
 
 export default class MessageReporter extends BaseReporter {
@@ -17,6 +18,7 @@ export default class MessageReporter extends BaseReporter {
     super(internalOptions);
     this.setOutputStream(this.userOptions.outputFile);
     this.eventBroadcaster.on('envelope', (envelope: messages.Envelope) => {
+      if (!isAttachmentAllowed(envelope, this.userOptions.skipAttachments)) return;
       this.outputStream.write(JSON.stringify(envelope) + '\n');
     });
   }